Decl impl2

+++ pythran/pythonic/include/__builtin__/id.hpp
@@ -1,28 +1,25 @@
 #ifndef PYTHONIC_INCLUDE_BUILTIN_ID_HPP
 #define PYTHONIC_INCLUDE_BUILTIN_ID_HPP
 
-#include <cstdint>
-#include "pythonic/types/long.hpp"
-#include "pythonic/utils/proxy.hpp"
+#include "pythonic/include/types/long.hpp"
+#include "pythonic/include/utils/proxy.hpp"
 
-namespace pythonic {
+namespace pythonic
+{
 
-    namespace __builtin__ {
+  namespace __builtin__
+  {
 
-        template <class T>
-            intptr_t id(T const & t);
+    template <class T>
+    long id(T const &t);
 
-        intptr_t id(pythran_long_t const &t);
-        intptr_t id(long const &t);
-        intptr_t id(double const &t);
-        intptr_t id(bool const &t);
+    long id(pythran_long_t const &t);
+    long id(long const &t);
+    long id(double const &t);
+    long id(bool const &t);
 
-        PROXY_DECL(pythonic::__builtin__, id);
-
-    }
-
+    PROXY_DECL(pythonic::__builtin__, id);
+  }
 }
 
 #endif
-
-
